// src/app/types/weather-gov/weatherGov.d.ts
export interface components {
  schemas: {
// ...
    AlertCollectionGeoJson: components['schemas']['GeoJsonFeatureCollection'] & {
      features?: {
        properties?: components['schemas']['Alert'];
      }[];
    } & components['schemas']['AlertCollection'];
// ...
  };
  responses: {
    /** @description A collection of alerts. */
    AlertCollection: {
      headers: {
        'X-Correlation-Id': components['headers']['CorrelationId'];
        'X-Request-Id': components['headers']['RequestId'];
        'X-Server-Id': components['headers']['ServerId'];
        [name: string]: unknown;
      };
      content: {
        'application/geo+json': components['schemas']['AlertCollectionGeoJson'];
        'application/ld+json': components['schemas']['AlertCollectionJsonLd'];
        'application/atom+xml': components['schemas']['AlertAtomFeed'];
      };
    };
    /** @description An error response. */
    Error: {
      headers: {
        'X-Correlation-Id': components['headers']['CorrelationId'];
        'X-Request-Id': components['headers']['RequestId'];
        'X-Server-Id': components['headers']['ServerId'];
        [name: string]: unknown;
      };
      content: {
        'application/problem+json': components['schemas']['ProblemDetail'];
      };
    };
    /** @description A forecast for a gridpoint. */
    GridpointForecast: {
      headers: {
        'X-Correlation-Id': components['headers']['CorrelationId'];
        'X-Request-Id': components['headers']['RequestId'];
        'X-Server-Id': components['headers']['ServerId'];
        [name: string]: unknown;
      };
      content: {
        'application/geo+json': components['schemas']['GridpointForecastGeoJson'];
        'application/ld+json': components['schemas']['GridpointForecastJsonLd'];
        'application/vnd.noaa.dwml+xml': unknown;
      };
    };
    /** @description An observation record. */
    Observation: {
      headers: {
        'X-Correlation-Id': components['headers']['CorrelationId'];
        'X-Request-Id': components['headers']['RequestId'];
        'X-Server-Id': components['headers']['ServerId'];
        [name: string]: unknown;
      };
      content: {
        'application/geo+json': components['schemas']['ObservationGeoJson'];
        'application/ld+json': components['schemas']['ObservationJsonLd'];
        'application/vnd.noaa.uswx+xml': unknown;
        'application/vnd.noaa.obs+xml': unknown;
      };
    };
    /** @description A collection of observation records. */
    ObservationCollection: {
      headers: {
        'X-Correlation-Id': components['headers']['CorrelationId'];
        'X-Request-Id': components['headers']['RequestId'];
        'X-Server-Id': components['headers']['ServerId'];
        [name: string]: unknown;
      };
      content: {
        'application/geo+json': components['schemas']['ObservationCollectionGeoJson'];
        'application/ld+json': components['schemas']['ObservationCollectionJsonLd'];
      };
    };
    /** @description A collection of observation stations. */
    ObservationStationCollection: {
      headers: {
        'X-Correlation-Id': components['headers']['CorrelationId'];
        'X-Request-Id': components['headers']['RequestId'];
        'X-Server-Id': components['headers']['ServerId'];
        [name: string]: unknown;
      };
      content: {
        'application/geo+json': components['schemas']['ObservationStationCollectionGeoJson'];
        'application/ld+json': components['schemas']['ObservationStationCollectionJsonLd'];
      };
    };
  };
// ...
}